+++ sdk-node/test/conformance/dispatch-features.ts
@@ -52,6 +52,7 @@ export function dispatchFeatures(
     case 'bill':
     case 'renew':
     case 'previewUpcomingInvoice':
+    case 'retrieveManageUrl':
     case 'count':
     case 'delete':
     case 'bulkUpsert':